refactor : creating separate file for schema validations
